pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coding-agent-plugin"  # Must be unique on PyPI
version = "0.1.0"
description = "Multi-agent coding assistant with orchestration, planning, and execution capabilities"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Mohamed Abu Basith", email = "your.email@example.com" }
]
maintainers = [
    { name = "Mohamed Abu Basith", email = "your.email@example.com" }
]
keywords = ["ai", "coding-agent", "llm", "multi-agent", "code-generation", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
    "minio>=7.2.0", # Optional MinIO support
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/coding-agent-plugin"
Documentation = "https://github.com/yourusername/coding-agent-plugin#readme"
Repository = "https://github.com/yourusername/coding-agent-plugin"
Issues = "https://github.com/yourusername/coding-agent-plugin/issues"

[project.scripts]
coding-agent = "coding_agent_plugin.cli.main:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/coding_agent_plugin"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "-s",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings"
]

